new `valence_instance`, reorganize `valence_client`
