+++ crates/valence_client/src/keepalive.rs
@@ -66,7 +66,7 @@ fn send_keepalive(
             } else {
                 let millis = settings.period.as_millis();
                 warn!("Client {entity:?} timed out: no keepalive response after {millis}ms");
-                commands.entity(entity).remove::<Client>();
+                commands.entity(entity).insert(Despawned);
             }
         }
     }
@@ -82,13 +82,13 @@ fn handle_keepalive_response(
             if let Ok((client, mut state, mut ping)) = clients.get_mut(packet.client) {
                 if state.got_keepalive {
                     warn!("unexpected keepalive from client {client:?}");
-                    commands.entity(client).remove::<Client>();
+                    commands.entity(entity).insert(Despawned);
                 } else if pkt.id != state.last_keepalive_id {
                     warn!(
                         "keepalive IDs don't match for client {client:?} (expected {}, got {})",
                         state.last_keepalive_id, pkt.id,
                     );
-                    commands.entity(client).remove::<Client>();
+                    commands.entity(entity).insert(Despawned);
                 } else {
                     state.got_keepalive = true;
                     ping.0 = state.last_send.elapsed().as_millis() as i32;
