+++ crates/valence_entity/src/packet.rs
@@ -197,6 +197,16 @@ pub struct EntitySpawnS2c {
     pub velocity: [i16; 3],
 }
 
+#[derive(Copy, Clone, Debug, Encode, Decode, Packet)]
+#[packet(id = packet_id::PLAYER_SPAWN_S2C)]
+pub struct PlayerSpawnS2c {
+    pub entity_id: VarInt,
+    pub player_uuid: Uuid,
+    pub position: DVec3,
+    pub yaw: ByteAngle,
+    pub pitch: ByteAngle,
+}
+
 #[derive(Clone, Debug, Encode, Decode, Packet)]
 #[packet(id = packet_id::ENTITY_STATUS_EFFECT_S2C)]
 pub struct EntityStatusEffectS2c {
