+++ crates/valence_instance/Cargo.toml
@@ -20,3 +20,4 @@ valence_dimension.workspace = true
 valence_entity.workspace = true
 valence_nbt.workspace = true
 valence_registry.workspace = true
+tracing.workspace = true
